+++ ricaun.Revit.UI.Example/Revit/App.cs
@@ -12,7 +12,8 @@ using System.Windows.Media;
 
 namespace ricaun.Revit.UI.Example.Revit
 {
-    //[AppLoader]
+    [AppLoader]
+    //[Obsolete]
     public class App : IExternalApplication
     {
         private const string TabName = "ricaun";